requirements.txt
# Core dependencies
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
starlette==0.40.0

# Database
SQLAlchemy==2.0.38
alembic==1.14.1
asyncpg==0.30.0
# psycopg2-binary==2.9.6 # Removed, incompatible with Python 3.13, using psycopg 3 instead
# psycopg==3.2.5 # Base package doesn't include necessary binary parts
psycopg[binary]==3.2.5 # Install with pre-compiled binary extras
greenlet==3.1.1

# HTTP and API
aiohttp==3.9.3
httpx==0.28.1
requests==2.32.3
python-multipart==0.0.6
googlemaps==4.10.0

# Authentication
PyJWT==2.10.1
python-jose==3.3.0
cryptography==44.0.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.2
orjson==3.10.15
tenacity==8.2.3
validators==0.20.0
email-validator==2.2.0

# Scraping
beautifulsoup4==4.13.3
lxml==5.2.2 # Updated for Python 3.13 compatibility
scraperapi-sdk==1.5.3

# Scheduling
APScheduler==3.10.4

# Other dependencies required by the application
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
ecdsa==0.19.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
packaging==23.2
propcache==0.3.0
pyasn1==0.4.8
pycparser==2.22
regex==2024.11.6
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zstandard==0.23.0

# Code Analysis Tools
importlab>=0.8.1 # Added for unused code detection
pydeps>=3.0.0 # Added for improved static analysis (replaces static_import_trace.py)
